+++ src/engine/eval/mod.rs
@@ -1,4 +1,5 @@
 mod material;
+mod mobility;
 mod params;
 mod phased_eval;
 pub mod piece_square_tables;
@@ -57,7 +58,9 @@ pub fn eval(game: &Game) -> Eval {
 }
 
 pub fn absolute_eval(game: &Game) -> WhiteEval {
-    let eval = game.incremental_eval.piece_square_tables + material::eval(game);
+    let eval = game.incremental_eval.piece_square_tables
+        + material::eval(game)
+        + mobility::eval(game);
 
     eval.for_phase(game.incremental_eval.phase_value)
 }